skills/serializer.py
from rest_framework import serializers
from .models import Skill

class SkillSerializer(serializers.ModelSerializer):
    class Meta:
        model = Skill
        fields = "__all__"

    def limit_queryset(self, queryset):
        return queryset[:100]